pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.20.3",
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.4",
    "langchain[anthropic,openrouter]>=0.3.25",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
]
